demo/garden/plant/pyproject.toml
[tool.poetry]
name = "plant"
version = "0.1.0"
description = ""
authors = [ "Linus Meierhoefer <linusmeierhoefer@protonmail.com>" ]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
p2pindustries = {path = "../../../sdks/python"}
grpcio = "^1.65.5"
protobuf = "^5.27.0"
pillow = "^10.4.0"
ltr559 = "^1.0.0"
st7735 = "^1.0.0"
fonts = "^0.0.3"
font-roboto = "^0.0.1"
gpiod = "^2.2.1"
gpiodevice = "^0.0.5"
gpiozero = "^2.0.1"
lgpio = "^0.2.2.0"
setuptools = "^73.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.5"

[tool.ruff]
indent-width = 4
target-version = 'py310'

[tool.ruff.format]
quote-style = 'single'
indent-style = 'space'
line-ending = 'auto'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
